Use redis to store the state of the components in a "session"
