+++ pyproject.toml
@@ -33,6 +33,7 @@ dependencies = [
    "orjson>=3.10.7",
    "channels>=4.1.0",
    "uuid6>=2024.7.10",
+   "redis[hiredis]>=5.0.8",
 ]
