+++ src/djhtmx/introspection.py
@@ -261,5 +261,9 @@ def is_simple_annotation(ann):
     return is_basic_type(ann) or is_union_of_basic(ann)
 
 
-Unset = object()
+class UnsetType:
+    pass
+
+
+Unset = UnsetType()
 _SIMPLE_TYPES = (int, str, float, UUID, types.NoneType, date, datetime, bool)
