+++ src/djhtmx/templates/htmx/headers.html
@@ -2,7 +2,8 @@
 
 <meta name="django-csrf-header-name" content="{{ CSRF_HEADER_NAME }}" />
 <meta name="django-csrf-token" content="{{ csrf_token }}" />
-<meta name="htmx-config" content='{"defaultSwapStyle": "outerHTML"}' />
+<meta name="hx-session" content="{{ hx_session }}" />
+<meta name="htmx-config" content='{"allowNestedOobSwaps": false}' />
 {% for script_url in SCRIPT_URLS %}
   <script defer src="{% static script_url %}"></script>
 {% endfor %}
