+++ src/djhtmx/utils.py
@@ -1,15 +1,44 @@
 import hashlib
 import typing as t
+from urllib.parse import urlparse
 
 from channels.db import database_sync_to_async as db  # type: ignore
 from django.db import models
+from django.http.request import HttpRequest, QueryDict
 from uuid6 import uuid7
+
+from . import json
 
 if t.TYPE_CHECKING:
     T = t.TypeVar("T")
     P = t.ParamSpec("P")
 
     def db(f: t.Callable[P, T]) -> t.Callable[P, t.Awaitable[T]]: ...
+
+
+def get_params(obj: HttpRequest | QueryDict | str | None) -> QueryDict:
+    if isinstance(obj, HttpRequest):
+        is_htmx_request = json.loads(obj.META.get("HTTP_HX_REQUEST", "false"))
+        if is_htmx_request:
+            return QueryDict(
+                urlparse(obj.META["HTTP_HX_CURRENT_URL"]).query,
+                mutable=True,
+            )
+        else:
+            qd = QueryDict(None, mutable=True)
+            qd.update(obj.GET)
+            return qd
+    elif isinstance(obj, QueryDict):
+        qd = QueryDict(None, mutable=True)
+        qd.update(obj)  # type: ignore
+        return qd
+    elif isinstance(obj, str):
+        return QueryDict(
+            query_string=urlparse(obj).query if obj else None,
+            mutable=True,
+        )
+    else:
+        return QueryDict(None, mutable=True)
 
 
 def get_model_subscriptions(
