+++ src/tests/fision/todo/templates/base.html
@@ -10,7 +10,7 @@
     <link rel="stylesheet" href="{% static "todo.css" %}" />
     {% htmx-headers %}
   </head>
-  <body hx-boost="true" hx-ext="ws" ws-connect="_htmx/ws">
+  <body hx-boost="true">
     <section>
       {% block body %}
       {% endblock body %}
