+++ src/tests/fision/todo/templates/todo/TodoList.html
@@ -13,7 +13,7 @@
           <!-- These are here just to show the structure of the list items -->
           <!-- List items should get the class `editing` when editing and `completed` when marked as completed -->
           {% for item in this.items %}
-            {% htmx 'TodoItem' id='item-id-'|concat:item.id item=item %}
+            {% htmx 'TodoItem' id='item-id-'|add:item.id item=item %}
           {% endfor %}
         </ul>
       {% endif %}
